UtilsKit.ForNumber

Exported

UtilsKit.ForNumber — Module

UtilsKit.ForNumber

Number utilities:

clamping and min/max helpers

invalid-number detection (nothing/missing/NaN/Inf)

invalid-value replacement and simple helpers like cumulative sum

UtilsKit.ForNumber.at_least_one — Method

at_least_one(num)

returns max(num, 1)

Examples

julia> using UtilsKit

julia> at_least_one(0.5)
1.0

UtilsKit.ForNumber.at_least_zero — Method

at_least_zero(num)

returns max(num, 0)

Examples

julia> using UtilsKit

julia> at_least_zero(-1.0)
0.0

UtilsKit.ForNumber.at_most_one — Method

at_most_one(num)

returns min(num, 1)

Examples

julia> using UtilsKit

julia> at_most_one(2.0)
1.0

UtilsKit.ForNumber.at_most_zero — Method

at_most_zero(num)

returns min(num, 0)

Examples

julia> using UtilsKit

julia> at_most_zero(1.0)
0.0

UtilsKit.ForNumber.clamp_zero_one — Method

clamp_zero_one(num)

returns max(min(num, 1), 0)

Examples

julia> using UtilsKit

julia> clamp_zero_one(2.0)
1.0

julia> clamp_zero_one(-0.5)
0.0

UtilsKit.ForNumber.cumulative_sum! — Method

cumulative_sum!(output::AbstractVector, input::AbstractVector)

fill out the output vector with the cumulative sum of elements from input vector

Examples

julia> using UtilsKit

julia> out = zeros(Int, 3);

julia> cumulative_sum!(out, [1, 2, 3])
3-element Vector{Int64}:
 1
 3
 6

UtilsKit.ForNumber.is_invalid_number — Method

is_invalid_number(_data::Number)

Checks if a number is invalid (e.g., nothing, missing, NaN, or Inf).

Arguments:

_data: The input number.

Returns:

true if the number is invalid, otherwise false.

Examples

julia> using UtilsKit

julia> is_invalid_number(NaN)
true

julia> is_invalid_number(1.0)
false

UtilsKit.ForNumber.replace_invalid_number — Method

replace_invalid_number(_data, _data_fill)

Replaces invalid numbers in the input with a specified fill value.

Arguments:

_data: The input number.

_data_fill: The value to replace invalid numbers with.

Returns:

The input number if valid, otherwise the fill value.

Examples

julia> using UtilsKit

julia> replace_invalid_number(NaN, 0.0)
0.0

julia> replace_invalid_number(2.0, 0.0)
2.0

UtilsKit.ForNumber.safe_divide — Method

safe_divide(num, den)

return either a ratio or numerator depending on whether denomitor is a zero

Examples

julia> using UtilsKit

julia> safe_divide(1.0, 2.0)
0.5

julia> safe_divide(1.0, 0.0)
1.0

Internal